+++ CHANGELOG.md
@@ -18,6 +18,7 @@ You can also check the
 - Fix
 
   - Map page: Shadow for tooltip
+  - Charts: Tooltip legend symbol does not get squished
 
 # 2.5.0 - 2025-07-09